pyproject.toml
[tool.poetry]
name = "pinta"
version = "0.0.1"
description = "Job management on GPU clusters"
authors = ["Pinta Team <pinta-l@usc.edu>"]
packages = [{ include = "pinta"}]
license = "AGPL-3.0"
readme = "README.md"
homepage = "https://qed.usc.edu/pinta"
repository = "https://github.com/qed-usc/pinta"
keywords = ["hpc", "gpu", "cluster", "scheduling"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Operating System :: OS Independent",
    "Topic :: System :: Clustering",
    "Topic :: System :: Distributed Computing"
]

[tool.poetry.dependencies]
python = "^3.8"
click = "^7.1.2"
httpx = "^0.13.3"
rich = "^2.0.1"
websockets = "^8.1"
requests = "^2.23.0"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pudb = "^2019.2"
pylint = "^2.5.3"
pep8 = "^1.7.1"
mkdocs-material = "^5.2.3"

[tool.poetry.scripts]
pinta = 'pinta.cli.__main__:main'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
